Move ipc classes to ipc module
