+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/ui/serviceconnection/VpnPermission.kt
@@ -1,9 +1,9 @@
 package net.mullvad.mullvadvpn.ui.serviceconnection
 
 import android.os.Messenger
-import net.mullvad.mullvadvpn.ipc.Event
-import net.mullvad.mullvadvpn.ipc.MessageDispatcher
-import net.mullvad.mullvadvpn.ipc.Request
+import net.mullvad.mullvadvpn.lib.ipc.Event
+import net.mullvad.mullvadvpn.lib.ipc.MessageDispatcher
+import net.mullvad.mullvadvpn.lib.ipc.Request
 
 class VpnPermission(private val connection: Messenger, eventDispatcher: MessageDispatcher<Event>) {
     var onRequest: (() -> Unit)? = null
